Converts Build Script to PowerShell
In a push to automate project builds, this changes the build script from a BAT script to a PowerShell script. This is a cross-platform solution that should allow us to build the project within the `mcr.microsoft.com/dotnet/sdk:6.0` docker image (Debian).

Also updates the README to list PowerShell 7 as a dependency as the `pwsh`  tool is not available in Windows PowerShell (v5).