fixed yet another method not being renamed properly LoadOfflineRaidScreenPatch for the "Ready" callback (!120)
When pressing ready on the scav offline raid screen the OnShowReadyScreen callback currently gets assigned to method_73. But that is the new name of the insurance screen as seen in 2abc1ab0ce7e6c8e7fe2365cae2321df2ccfa4bc.
And this is indeed what currently happens. You get your PMC insurance screen.
The correct method name it should register appears to be method_76. My reasoning for this is that method_44, which is the one called for non-pmcs in method_72, registers method_76 as the ready callback in the original unpatched code.

I am trying to be less gung ho and have checked more thoroughly this time, but please tripple check as I've fucked up once already!

Reviewed-on: https://dev.sp-tarkov.com/SPT-AKI/Modules/pulls/120
Co-authored-by: BluBb_mADe <ku21runy1k11@opayq.com>
Co-committed-by: BluBb_mADe <ku21runy1k11@opayq.com>