Add GUI Component to Bundle Loading (!140)
This gives users feedback that the game is currently loading bundles, which prevents confusion when a large amount of them are being loaded and the game is "seemingly" infinitely loading.
The implementation is very bare bones, feel free to modify it to your liking if you want it to use a more fancy user interface. I went for clarity (black background, white text). With the standard `OnGUI` skin it's very hard to discern the text.
Tested on 1440p and 1080p.

![image](/attachments/31a818d6-d101-4e5d-a2c5-541c8b48ba71)

Co-authored-by: Lacyway <20912169+Lacyway@users.noreply.github.com>
Reviewed-on: https://dev.sp-tarkov.com/SPT/Modules/pulls/140
Co-authored-by: Lacyway <lacyway@noreply.dev.sp-tarkov.com>
Co-committed-by: Lacyway <lacyway@noreply.dev.sp-tarkov.com>