Remove bundle before re-acquiring (!155)
Currently, the system knows which bundles are incorrect and attempts to re-download the correct ones.. However, upon next re-launch those bundles will again be invalid probably due to it not actually writing to them forcing the system to keep re-downloading.

This fixes it by removing the invalid bundle, then re-downloading the proper one.

Reviewed-on: https://dev.sp-tarkov.com/SPT/Modules/pulls/155
Co-authored-by: Archangel <jesse@archangel.wtf>
Co-committed-by: Archangel <jesse@archangel.wtf>