Rename ShouldReaquire & check if bundle exists (!160)
I didn't keep in mind that 'ShouldReaquire' handles both the initial check for acquiring of a file and also the re-acquiring of it, I've renamed the method here to be more accurate in this behavior so that hopefully it prevents this from happening again in the future.

Also fixes the issue of an exception being thrown when checking if the bundle exists or not.

Thanks to MadByte for finding this issue

Reviewed-on: https://dev.sp-tarkov.com/SPT/Modules/pulls/160
Co-authored-by: Archangel <jesse@archangel.wtf>
Co-committed-by: Archangel <jesse@archangel.wtf>