Rework RemoveStashUpgradeLabelPatch (!162)
The `InventoryScreen` is part of the `CommonUI` which lives throughout the lifetime of the application. It should be safe to only delete this on `Awake()` rather than use `Find()` on a transform that has a fair bit of children every time the inventory screen is opened.
![image](/attachments/937a3954-210c-4fc7-a869-75458ae19d8d)
I cannot extract on 3.10 currently so need someone to just verify that it sticks after an extract where the menu is reloaded.

Co-authored-by: Lacyway <20912169+Lacyway@users.noreply.github.com>
Reviewed-on: https://dev.sp-tarkov.com/SPT/Modules/pulls/162
Co-authored-by: Lacyway <lacyway@noreply.dev.sp-tarkov.com>
Co-committed-by: Lacyway <lacyway@noreply.dev.sp-tarkov.com>