310-dev (!168)
This adds a patch that enforces PMC AI to be hostile towards scavs. From my testing they were not attacking random scavs and the player because the boolean failed at the attached screenshot due to PMCs being bosses and the loyalty check. Now it checks if the initial bot type of the group was a PMC Bear or USEC and then if the player it checks for is a scav, they will be forced to be an enemy (which to me seemed natural).
![image](/attachments/4a73df0f-7030-4c61-869f-0a20ba804d8b)

I've tested this on 3 raids on customs and it seems to work fine. Feel free to edit the PR if the logic is flawed, and please test the patch thoroughly to ensure my tests were not just lucky.
Thanks!

Co-authored-by: Lacyway <20912169+Lacyway@users.noreply.github.com>
Reviewed-on: https://dev.sp-tarkov.com/SPT/Modules/pulls/168
Co-authored-by: Lacyway <lacyway@noreply.dev.sp-tarkov.com>
Co-committed-by: Lacyway <lacyway@noreply.dev.sp-tarkov.com>