Fix scav exfils (!169)
`InitAllExfiltrationPoints` will instantiate the list of `ScavExfiltrationPoints` in one line before `EligiblePoints` is ran, meaning the following always returned:
```cs
            if (__instance.ScavExfiltrationPoints.Length > 0)
            {
                Logger.LogError($"ScavExfiltrationPoints has content, Do original");
                foreach (var scavExit in __instance.ScavExfiltrationPoints)
                {
                    Logger.LogError($"{scavExit.name}, {scavExit.Id}, {scavExit.Description}");
                }

                return true; // do original
            }
```

I'm not sure if the logic was different before 3.10, but removing this ensures that the patch runs as intended. Tested both PMC and Scav raids, all is fine for me. Please test on your end as well.

Co-authored-by: Lacyway <20912169+Lacyway@users.noreply.github.com>
Reviewed-on: https://dev.sp-tarkov.com/SPT/Modules/pulls/169
Co-authored-by: Lacyway <lacyway@noreply.dev.sp-tarkov.com>
Co-committed-by: Lacyway <lacyway@noreply.dev.sp-tarkov.com>