Fix FixSavageInventoryScreenPatch (!173)
This fixes the logic of FixSavageInventoryScreenPatch.
NOTE:  The code below is untested and will most likely not work. It's what I theoretically believe is the right way to turn the `JObject` into a string for the server, however BSG has a `JsonWriter` in the client now that might have to be utilized? I will leave it like this until we have more info.
```c#
RequestHandler.PutJson("/raid/profile/scavsave",
				GetProfileAtEndOfRaidPatch.ProfileDescriptor.ToUnparsedData([]).JObject.ToString());
```

Co-authored-by: Lacyway <20912169+Lacyway@users.noreply.github.com>
Reviewed-on: https://dev.sp-tarkov.com/SPT/Modules/pulls/173
Co-authored-by: Lacyway <lacyway@noreply.dev.sp-tarkov.com>
Co-committed-by: Lacyway <lacyway@noreply.dev.sp-tarkov.com>